[AutoDiff] Revert obsolete SIL undef hack.
Previously, JVP/VJP generation used a "return undef" hack when
differential/pullback values did not match the expected return type.

This was relevant before differentiation supported "loadable types with
address-only tangent types", which was diagnosed.

Now that support for the above exists, "return undef" should be removed and
replaced with an assertion.