[TypeChecker/CodeCompletion] Make viability checking responsibility of `typeCheckForCodeCompletion`
There is no way to separate viability/applicability checking from
`TypeChecker::typeCheckForCodeCompletion` because multi-statement
closures could be either type-checked together with enclosing
context (e.g. when closure represents a function builder body), or
separately - when it's just a regular closure.

Due to this "duality" we need to attempt to run code completion
code to determine whether body of the closure participated in
type-check and if it didn't, fallback to a `typeCheckExpression`.