Fix ConnectionGraph verification for calls to no return functions.
Functions that do not have a return, and instead end with 'unreachable'
due to NoReturnFolding will not have a ReturnNode in the connection
graph.

A caller calling a no return function then may not have a CGNode
corresponding to the call's result.

Fix the ConnectionGraph's verifier so we don't assert in such cases.