[ownership] Merge regularUsers and implicitRegularUsers into just "regularUsers".
When bringing up ownership I thought that it might be useful to distinguish
"normal users" that just require liveness and are real transitive users vs
implicit users that require liveness (I provide a constrasting example
below). Turns out this distinction did not provide any benefit, so I am ripping
it out so I can refactor this code to be used in other parts of the compiler
where we only want a single array of "normal users".

This is just a pure refactor. NFCI.

--------------------------------------------------------------------------------

An example of the former would be an apply that uses an owned value as a
guaranteed parameter:

```
bb0(%0 : @owned $Klass):
  apply %func(%0) : $@convention(thin) (@guaranteed $Klass) -> ()
```

while an example of the latter is an end_apply of a coroutine that takes an
owned value as a guaranteed parameter:

```
bb0(%0 : @owned $Klass):
  (%result, %token) = begin_apply %func(%0)
  ...
  end_apply %token
  destroy_value %0
```

In the latter, we can not move the destroy_value past the end_apply.