Generalize the "semantic members" sort to all implicit members.
The "semantic members" query does a sort of synthesized functions to
ensure that they are provided in a stable order across translation
units. Generalize this sort operation to all implicit members, so that
`getSemanticMembers()` is deterministic.

While here, optimize the sort a bit so we only compute interface types
and turn them into strings in cases where the declaration type and
name are equivalent.