Replace initializedness-tracking approach with an approach based on visiting all (un)initializing instructions where at least one parameter is active
We found out that it is very untrivial to correctly check initializedness of code with aggregate projections
(e.g. nested structs with multiple `struct_element_addr` referring to the same element). Instead of checking
initializedness, we can rely on correctness of the original function SIL and reproduce (un)initializing
instructions based on the original SIL.