SILOptimizer: make vtable pruning less aggressive
A class which is marked as internal or public can be visible outside
of the current file, where the use of the VWT indirectly is possible.
If the VWT is modified and inlined, it is possible that the offsets will
no longer match resulting in an invalid dispatch.  Limit the pass to
when WMO is enabled or the type is private in non-WMO cases.