[AutoDiff] [Sema] Fix missing access-level mismatch diagnostic in `-enable-testing` builds.
When a derivative function is internal and its original function is
public, there should be an error when the derivative is not `@usableFromInline`.
This patch fixes a bug where the error does not appear in SwiftPM debug
build configurations (or any configurations with `-enable-testing`).

`ValueDecl::getEffectiveAccess()` should not be used when we perform
formal access checking for AutoDiff-related declaration attributes because
it will treat all internal declarations as public when `-enable-testing` is
enabled. Instea we should use a combination of `ValueDecl::getFormalAccess()`
and `ValueDecl::isUsableFromInline()`.

Also, add a `RUN` line with `-enable-testing` to all AutoDiff declaration
attribute test files.

Resolves SR-13500 (rdar://68336300).