[Code completion] Update the set of parsed members for delayed declarations.
Code completion delays the parsing of the declaration in which code
completion is happening, and then inserts it later. This happens
*after* we have populated the parse-members request, breaking the
model that this result is functional and immutable.

Patch up the result of the parse-members request when this happens, so
the rest of the frontend sees the correct (completed) result. This is a
localized hack to a more general problem: we shouldn't ever be
populating the parse-members request without including the delayed
declaration in it. However, addressing that is a more significant
refactoring.