[WIP] Add ThrowsType member to AnyFunctionType
This commit is quite big, because the constructor of AnyThrowsType and
of its subclasses is used extremely much throughout the whole codebase.
Although it technically compiles, it is not building the stdlib, as I'm
getting lots of these failed assertions:

Assertion failed: (repr && "Cannot validate null TypeReprs!")
in function resolveType at lib/Sema/TypeCheckType.cpp:1872.

I suspect, that these errors are coming from a mistake, I made in the
serialization or deserialization of function types and/or declarations.