Sema: Tweak pre-check unqualified lookup behavior to match parse-time lookup
preCheckExpression() resolves any DeclRefExprs that were not resolved
by parse-time lookup using ASTScope lookup.

In preparation for disabling parse-time lookup, tweak pre-check lookup
to match the behavior of parse-time lookup when a variable references
another variable with the same name inside its initializer, eg:

  func foo(x: Int) {
    let x = x
  }

We want to consider the usage of 'x' in the initializer to come "before"
'let x', so we use the end of the source range of the initializer instead
of the source location of the property (which points at the name 'x'),
and change the less-than check to a less-than-or-equal-to.