[CI] Re-enable test
Disable for "remote runs".  We don't copy the corresponding `.dSYM`
with the test binary, and hence symbolication (source info) does not
succeed.

Also explicitly mark this test unsupported on for the 32-bit iOS
simulator.  It has an outdated version of `atos` that can't demangle
Swift symbol names.

Enable on other platforms.

rdar://68353068